Move methods of uploads into controller
